Search by name using event query API
